--- src/page/Todo/style.jsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const Wrapper = styled.div`
-  width: 100%;
-  position: absolute;
-  top: 10%;
-  padding: 0 32px;
-`;
